fix(products): nine defects the benidorm stand found, and P-D-164's record of the wave
The gear's own tier runs against an authz double and SQLite; every defect below
lives where that double ends. The e2e suite in vhp-core (`bss/products-e2e`, 98
scenarios, five deploys) is the second reading, and this is what it caught.

The one that mattered: `products_approval` declares `resource_col = "approval_id"`,
and every head door read its gate candidates under its own AccessScope — which a
real PDP compiles with `resource_id` pinned to the ENTITY, so the secure layer
filtered `approval_id = <entity id>` and no record could ever match. Correct,
deprecate, undeprecate, retire and every governed live op answered 403
APPROVAL_REQUIRED beside their satisfied records; `supersede_open_approval` was
the quiet half, matching no row and calling it success. The four approval-store
functions scope by tenant now.

Also: the gear's init-order dependencies (the collector client was read before
the collector initialised); `generation` widened to bigint on both read tables
(an i64 field over an INT4 column 500'd every browse after a checkpoint, with a
width oracle to pin it); the actor-ref mint race and its mirror image, an
out-of-order `last_seen_at` that broke the seen-order CHECK; the projection
carrying the head's own `sellable` instead of deriving one; and the allow-list
door owning its refusals — an omitted `signedOffBy` rides VALIDATION naming the
field (P-D-64) instead of serde's bare 422, and a duplicate value answers
DUPLICATE_CODE where the partial index used to surface a 500. Two probes for the
last pair in the gear's tier.

Named rather than fixed, in P-D-164: the decide door's missing role claim and
break-glass roles at the substituted tenant (both platform), the governance
bootstrap on a fresh tenant, the SQLite/Postgres split on a duplicate allow-list
value, and two contract gaps (compositionPending absent from the REST head view,
camelCase query keys silently ignored).

Gates: fmt, clippy -D warnings, doc, 1249 lib + 3 harness + 7 seam, SDK 23,
Postgres tier 54/54, cfs 0 errors / 20 warnings, spec-check 37 (baseline),
tick census 0 shadowed.

+++ gears/bss/products/products/src/infra/storage/migrations/m20260901_000023_create_products_read_entity.rs
@@ -94,7 +94,7 @@ const PG_UP_STATEMENTS: &[&str] = &[
             category_paths     text,
             published_version  bigint      NOT NULL,
             projected_at       timestamptz NOT NULL,
-            generation         integer     NOT NULL DEFAULT 0,
+            generation         bigint    NOT NULL DEFAULT 0,
             CONSTRAINT products_read_entity_pkey PRIMARY KEY (tenant_id, entity_kind, entity_id),
             CONSTRAINT chk_products_read_entity_kind CHECK (entity_kind IN ('product', 'sku')),
             CONSTRAINT chk_products_read_entity_state CHECK (lifecycle_state IN ('draft', 'published', 'deprecated', 'retired', 'discarded')),
